Merge release 1.3.0 into master

+++ jpo-ode-plugins/src/main/java/us/dot/its/jpo/ode/plugin/j2735/J2735RegulatorySpeedLimit.java
@@ -1,8 +1,10 @@
 package us.dot.its.jpo.ode.plugin.j2735;
 
+import java.math.BigDecimal;
+
 public class J2735RegulatorySpeedLimit {
 	private J2735SpeedLimitType type;
-	private Integer speed;
+	private BigDecimal speed;
 
 	public J2735SpeedLimitType getType() {
 		return type;
@@ -12,11 +14,11 @@ public class J2735RegulatorySpeedLimit {
 		this.type = type;
 	}
 
-	public Integer getSpeed() {
+	public BigDecimal getSpeed() {
 		return speed;
 	}
 
-	public void setSpeed(Integer speed) {
+	public void setSpeed(BigDecimal speed) {
 		this.speed = speed;
 	}
